controledecontratos/backend/requirements-lite.txt
Flask==3.0.3
Flask-Cors==4.0.1
SQLAlchemy==2.0.32
psycopg2-binary==2.9.11
python-dotenv==1.0.1
gunicorn==22.0.0
APScheduler==3.10.4
PyJWT==2.9.0
Werkzeug==3.0.3
